// modules/accounts-management/modules/ach/enum/url-collection.enum.ts
export enum AMAchUrlCollection {
  HOME = 'accounts-management/ach',
  CREATE = 'accounts-management/ach/create',
  CREATE_CONFIRMATION = 'accounts-management/ach/create/confirmation',
  CREATE_VOUCHER = 'accounts-management/ach/create/voucher',
  UPDATE = 'accounts-management/ach/update',
  UPDATE_CONFIRMATION = 'accounts-management/ach/update/confirmation',
  UPDATE_VOUCHER = 'accounts-management/ach/update/voucher',
  DELETE = 'accounts-management/ach/delete',
  DELETE_CONFIRMATION = 'accounts-management/ach/delete/confirmation',
  DELETE_VOUCHER = 'accounts-management/ach/delete/voucher',
  DETAIL = 'accounts-management/ach/detail',
}