// modules/accounts-management/modules/ach/enum/route-protected.enum.ts
export enum AmACHRouteProtected {
  DETAIL = 'ach:detail',
  CREATE = 'ach:create',
  CREATE_CONFIRM = 'ach:create_confirm',
  CREATE_VOUCHER = 'ach:create_voucher',
  UPDATE = 'ach:update',
  UPDATE_CONFIRM = 'ach:update_confirm',
  UPDATE_VOUCHER = 'ach:update_voucher',
  REMOVE = 'ach:delete',
  REMOVE_CONFIRM = 'ach:delete_confirm',
  REMOVE_VOUCHER = 'ach:delete_voucher',
}